move IdentityPublicKeyInCreationWASM to IdentityPublicKeyWASM package, validate_private_key
