+++ packages/identity_transitions/Cargo.toml
@@ -13,7 +13,6 @@ dpp = { workspace = true, features = [
 ]}
 js-sys = { workspace = true }
 pshenmic_dpp_utils = {path = "../utils"}
-pshenmic_dpp_contract_bounds = {path = "../contract_bounds"}
 pshenmic_dpp_core_script = {path = "../core_script"}
 pshenmic_dpp_identity_public_key = {path = "../identity_public_key" }
 pshenmic_dpp_identifier = {path = "../identifier" }
