+++ packages/identity_transitions/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod create_transition;
 pub mod credit_withdrawal_transition;
 pub mod identity_credit_transfer_transition;
-pub mod public_key_in_creation;
 pub mod top_up_transition;
 pub mod update_transition;
